Remove deferred rendering and layers.
To more closely match how Firefox renders we need to do this. These options are located in some magic checkboxes in xcode. In MainMenu.xib under Window look on the options panel on the right, one of them has deferred under it, uncheck that. Under View on the right again one of the panels has layers, uncheck the box beside View to remove the CoreAnimation layer from it.