Use undocumented _shouldRoundCornersForSurface to control the OS's clipping of our corners.
For some reason this only seems to have any effect when the opengl context is opaque. When the opengl context is not opaque we get the nice antialiased corners from the OS no matter what we return from _shouldRoundCornersForSurface. When the opengl context is opaque we either get the ugly non-antialiased rounded corners or a square window depending on what we return from _shouldRoundCornersForSurface.