Add an option to Draw via opengl into an IOSurface which is linked to a CoreAnimation Layer.
The exact same opengl drawing code is used the only difference is binding a framebuffer or not during it. The framebuffer is linked to the IOSurface. Note that binding a framebuffer number of 0 does nothing, so the opengl code doesn't draw into a framebuffer by default even though we call glBindFramebuffer.